[FIX] product, website_sale: partner pricelist fallback
When no pricelist property is defined for a customer, we use a search as 
fallback, without considering the current companies.

This means you can see the pricelist of company B as pricelist for a 
partner, even with company A as main current company (if both company 
are enabled in the multi-company widget).

closes odoo/odoo#46429

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>