[FIX] l10n_fr_pos_cert,point_of_sale: fix reward for FR pos
Install point_of_sale,l10n_fr_pos_cert,contacts,pos_loyalty
Create French company, add french chart of account
Create DEMO customer loaded with reward points
Enable loyalty program on POS, change the default reward to be:
* Reward Type: Discount(in value)
* Discount Product: create one (all default is ok)

Open POS, add a product, add the customer and the reward.

When you try to use the reward it is positive. You also get a "Decrease
the quantity by" pop-up when trying to remove it.
Since the reward setup conflict with the pos modification the fix is to
add an exception for the rewards

opw-2224086

closes odoo/odoo#49130

Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>