[FIX] account: fix invoice analysis report
In case of grouping by month, the graph should display a value
for every month. However, having months without invoices
are returning false value for the currency_id and subtotal which
trigger a traceback:
bool is not subscriptable on the check of the currency_id
Due to the fact that there is no currency on the current month.

fix commit 5e9f549331096f193dcba37f70f27b19e424888b
related task 2206699
opw-2232468

closes odoo/odoo#49229

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>