[FIX] account: fix multicompany journal audit report
With a second company generate the report under
Accounting>Reporting>Journals Audit
The name on the report and the currency are the one of the first
company.

This occur because the information on the template is gathered using a
parameter named 'res_company' which is set to the first company and not
the actual one (stored in company_id). Changing the parameter on the
template fix the issue.

opw-2242431

closes odoo/odoo#50098

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>